+++ ticketutil/redmine.py
@@ -1,21 +1,10 @@
 import logging
-import os
 
 import requests
 
 from . import ticket
 
 __author__ = 'dranck, rnester, kshirsal'
-
-# Disable warnings for requests because we aren't doing certificate verification
-requests.packages.urllib3.disable_warnings()
-
-DEBUG = os.environ.get('TICKETUTIL_DEBUG', 'False')
-
-if DEBUG == 'True':
-    logging.basicConfig(level=logging.DEBUG)
-else:
-    logging.basicConfig(level=logging.INFO)
 
 
 class RedmineTicket(ticket.Ticket):
@@ -31,7 +20,7 @@ class RedmineTicket(ticket.Ticket):
         # Redmine URLs
         self.url = url
         self.rest_url = '{0}/issues'.format(self.url)
-        self.auth_url = '{0}/projects/{1}.json'.format(self.url, project)
+        self.auth_url = '{0}/projects.json'.format(self.url)
 
         # Call our parent class's init method which creates our requests session.
         super(RedmineTicket, self).__init__(project, ticket_id)
@@ -51,6 +40,38 @@ class RedmineTicket(ticket.Ticket):
             ticket_url = "{0}/{1}".format(self.rest_url, self.ticket_id)
 
         return ticket_url
+
+    def _verify_project(self, project):
+        """
+        Queries the Redmine API to see if project is a valid project for the given Redmine instance.
+        :param project: The project you're verifying.
+        :return: True or False depending on if project is valid.
+        """
+        try:
+            r = self.s.get("{0}/projects/{1}.json".format(self.url, project))
+            logging.debug("Verify project: Status Code: {0}".format(r.status_code))
+            r.raise_for_status()
+            logging.debug("Project {0} is valid".format(project))
+            return True
+        except requests.RequestException as e:
+            logging.error("Project {0} is not valid.".format(project))
+            return False
+
+    def _verify_ticket_id(self, ticket_id):
+        """
+        Queries the Redmine API to see if ticket_id is a valid ticket for the given Redmine instance.
+        :param ticket_id: The ticket you're verifying.
+        :return: True or False depending on if ticket is valid.
+        """
+        try:
+            r = self.s.get("{0}/{1}.json".format(self.rest_url, ticket_id))
+            logging.debug("Verify ticket_id: Status Code: {0}".format(r.status_code))
+            r.raise_for_status()
+            logging.debug("Ticket {0} is valid".format(ticket_id))
+            return True
+        except requests.RequestException as e:
+            logging.error("Ticket {0} is not valid.".format(ticket_id))
+            return False
 
     def create(self, subject, description, **kwargs):
         """
@@ -266,6 +287,61 @@ class RedmineTicket(ticket.Ticket):
             except requests.RequestException as e:
                 logging.error("Error adding {0} as a watcher to ticket".format(watcher))
                 logging.error(e.args[0])
+
+    def add_attachment(self, file_name):
+        """
+        Attaches a file to a Redmine ticket.
+        :param file_name: A string representing the file to attach.
+        :return:
+        """
+        if not self.ticket_id:
+            logging.error("No ticket ID associated with ticket object. Set ticket ID with set_ticket_id(ticket_id)")
+            return
+
+        # First, upload the file to Redmine and retrieve a token to be used in subsequent request.
+        token = self._upload_file(file_name)
+
+        if token:
+            params = {'issue': {}}
+            params['issue']['uploads'] = [{'token': token, 'filename': file_name}]
+
+            try:
+                r = self.s.put('{0}/{1}.json'.format(self.rest_url, self.ticket_id), json=params)
+                r.raise_for_status()
+                logging.debug("Add attachment: Status Code: {0}".format(r.status_code))
+                logging.info("Attached file {0}: {1} - {2}".format(file_name, self.ticket_id, self.ticket_url))
+            except requests.RequestException as e:
+                logging.error("Error attaching file {0}".format(file_name))
+                logging.error(e.args[0])
+            except IOError:
+                logging.error("{0} not found".format(file_name))
+
+    def _upload_file(self, file_name):
+        """
+        Uploads a file to /uploads.json.
+        :param file_name: A string representing the file to upload.
+        :return: token: A token to be used in the request to add attachment.
+        """
+        headers = {'Content-Type': 'application/octet-stream'}
+        token = ''
+
+        # Upload file to uploads.json and retrieve token to be used in the add_attachment() request.
+        try:
+            params = open(file_name, 'rb')
+            r = self.s.post("{0}/uploads.json".format(self.url),
+                            data=params,
+                            headers=headers)
+            r.raise_for_status()
+            token = r.json()['upload']['token']
+            logging.debug("Upload attachment: Status Code: {0}".format(r.status_code))
+            logging.info("Uploaded file {0} to Redmine".format(file_name))
+        except requests.RequestException as e:
+            logging.error("Error uploading file {0}".format(file_name))
+            logging.error(e.args[0])
+        except IOError:
+            logging.error("{0} not found".format(file_name))
+
+        return token
 
     def _get_project_id(self):
         """