+++ ticketutil/rt.py
@@ -1,5 +1,4 @@
 import logging
-import os
 import re
 
 import requests
@@ -7,16 +6,6 @@ import requests
 from . import ticket
 
 __author__ = 'dranck, rnester, kshirsal'
-
-# Disable warnings for requests because we aren't doing certificate verification
-requests.packages.urllib3.disable_warnings()
-
-DEBUG = os.environ.get('TICKETUTIL_DEBUG', 'False')
-
-if DEBUG == 'True':
-    logging.basicConfig(level=logging.DEBUG)
-else:
-    logging.basicConfig(level=logging.INFO)
 
 
 class RTTicket(ticket.Ticket):
@@ -49,6 +38,52 @@ class RTTicket(ticket.Ticket):
             ticket_url = "{0}/Ticket/Display.html?id={1}".format(self.url, self.ticket_id)
 
         return ticket_url
+
+    def _verify_project(self, project):
+        """
+        Queries the RT API to see if project is a valid project for the given RT instance.
+        :param project: The project you're verifying.
+        :return: True or False depending on if project is valid.
+        """
+        try:
+            r = self.s.get("{0}/queue/{1}".format(self.rest_url, project))
+            logging.debug("Verify project: Status Code: {0}".format(r.status_code))
+            r.raise_for_status()
+            error_response = "No queue named {0} exists.".format(project)
+            # RT's API returns 200 even if the project is not valid. We need to parse the response.
+            if error_response in r.text:
+                logging.error("Project {0} is not valid.".format(project))
+                return False
+            else:
+                logging.debug("Project {0} is valid".format(project))
+                return True
+        except requests.RequestException as e:
+            logging.error("Unexpected error occurred when verifying project.")
+            logging.error(e.args[0])
+            return False
+
+    def _verify_ticket_id(self, ticket_id):
+        """
+        Queries the RT API to see if ticket_id is a valid ticket for the given RT instance.
+        :param ticket_id: The ticket you're verifying.
+        :return: True or False depending on if ticket is valid.
+        """
+        try:
+            r = self.s.get("{0}/ticket/{1}/show".format(self.rest_url, ticket_id))
+            logging.debug("Verify ticket_id: Status Code: {0}".format(r.status_code))
+            r.raise_for_status()
+            error_response = "Ticket {0} does not exist.".format(ticket_id)
+            # RT's API returns 200 even if the ticket is not valid. We need to parse the response.
+            if error_response in r.text:
+                logging.error("Ticket {0} is not valid.".format(ticket_id))
+                return False
+            else:
+                logging.debug("Ticket {0} is valid".format(ticket_id))
+                return True
+        except requests.RequestException as e:
+            logging.error("Unexpected error occurred when verifying ticket_id.")
+            logging.error(e.args[0])
+            return False
 
     def create(self, subject, text, **kwargs):
         """
@@ -90,22 +125,24 @@ class RTTicket(ticket.Ticket):
         :param subject: The ticket subject.
         :param text: The ticket text.
         :param fields: Other ticket fields.
-        :return: params: A string to pass in to the POST request containing ticket details.
+        :return: params: A dict to pass in to the POST request containing ticket details.
         """
-        # Weird format for request data that RT is expecting...
         # RT requires a special encoding on the Text parameter.
-        params = 'content='
-        params += 'Queue: {0}\n'.format(self.project)
-        params += 'Requestor: {0}\n'.format(self.principal)
-        params += 'Subject: {0}\n'.format(subject)
-        params += 'Text: {0}\n'.format(_text_encode(text))
+        encoded_text = text.replace('\n', '\n      ')
+
+        content = 'Queue: {0}\n'.format(self.project)
+        content += 'Requestor: {0}\n'.format(self.principal)
+        content += 'Subject: {0}\n'.format(subject)
+        content += 'Text: {0}\n'.format(encoded_text)
 
         # Some of the ticket fields need to be in a specific form for the tool.
         fields = _prepare_ticket_fields(fields)
 
         # Iterate through our options and add them to the params dict.
         for key, value in fields.items():
-            params += '{0}: {1}\n'.format(key.title(), value)
+            content += '{0}: {1}\n'.format(key.title(), value)
+
+        params = {'content': content}
 
         return params
 
@@ -151,11 +188,13 @@ class RTTicket(ticket.Ticket):
         # Some of the ticket fields need to be in a specific form for the tool.
         fields = _prepare_ticket_fields(kwargs)
 
-        params = 'content='
+        content = ''
 
-        # Iterate through our kwargs and add them to the params dict.
+        # Iterate through our kwargs and add them to the content string.
         for key, value in fields.items():
-            params += '{0}: {1}\n'.format(key.title(), value)
+            content += '{0}: {1}\n'.format(key.title(), value)
+
+        params = {'content': content}
 
         # Attempt to edit ticket.
         try:
@@ -177,9 +216,13 @@ class RTTicket(ticket.Ticket):
             logging.error("No ticket ID associated with ticket object. Set ticket ID with set_ticket_id(ticket_id)")
             return
 
-        params = 'content='
-        params += 'Action: correspond\n'
-        params += 'Text: {0}\n'.format(comment)
+        # RT requires a special encoding on the comment parameter.
+        encoded_comment = comment.replace('\n', '\n      ')
+
+        content = 'Action: correspond\n'
+        content += 'Text: {0}\n'.format(encoded_comment)
+
+        params = {'content': content}
 
         # Attempt to add comment to ticket.
         try:
@@ -201,8 +244,9 @@ class RTTicket(ticket.Ticket):
             logging.error("No ticket ID associated with ticket object. Set ticket ID with set_ticket_id(ticket_id)")
             return
 
-        params = 'content='
-        params += 'Status: {0}\n'.format(status.lower())
+        content = 'Status: {0}\n'.format(status.lower())
+
+        params = {'content': content}
 
         # Attempt to change status of ticket.
         try:
@@ -214,17 +258,33 @@ class RTTicket(ticket.Ticket):
             logging.error("Error changing status of ticket")
             logging.error(e.args[0])
 
+    def add_attachment(self, file_name):
+        """
+        Attaches a file to a RT ticket.
+        :param file_name: A string representing the file to attach.
+        :return:
+        """
+        if not self.ticket_id:
+            logging.error("No ticket ID associated with ticket object. Set ticket ID with set_ticket_id(ticket_id)")
+            return
 
-def _text_encode(text):
-    """
-    Encodes the text parameter in the form expected by RT's REST API when creating a ticket.
-    Replaces spaces with '+' and new lines with '%0A+'.
-    :param text: The string that will be used in the text field for our ticket.
-    :return: The string containing the replaced characters.
-    """
-    text = text.replace(' ', '+')
-    text = text.replace('\n', '%0A+')
-    return text
+        content = 'Action: correspond\n'
+        content += 'Attachment: {0}\n'.format(file_name)
+
+        params = {'content': content}
+        files = {'attachment_1': open(file_name, 'rb')}
+
+        # Attempt to attach file.
+        try:
+            r = self.s.post("{0}/ticket/{1}/comment".format(self.rest_url, self.ticket_id), data=params, files=files)
+            r.raise_for_status()
+            logging.debug("Add attachment: Status Code: {0}".format(r.status_code))
+            logging.info("Attached File {0}: {1} - {2}".format(file_name, self.ticket_id, self.ticket_url))
+        except requests.RequestException as e:
+            logging.error("Error attaching file {0}".format(file_name))
+            logging.error(e.args[0])
+        except IOError:
+            logging.error("{0} not found".format(file_name))
 
 
 def _prepare_ticket_fields(fields):